+++ src/pages/app/project-details.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import Header from '../../components/app/Header.tsx';
+import Header from '../../components/app/header';
 import SideBar from '../../components/app/SideBar.tsx';
 import PageContent from '../../components/app/PageContent.tsx';
 ---
